backend/requirements.txt
flask
flask-cors
python-dotenv
flask_sqlalchemy
psycopg2-binary
flask_migrate
pyyaml
redis
flask-socketio
eventlet
flask_jwt_extended